requirements.txt
customtkinter==5.2.2
numpy==1.26.4
requests==2.31.0
Pillow==10.2.0
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
PyPDF2==3.0.1
scikit-learn==1.4.1
pytesseract==0.3.10
flask==3.0.2